requirements.txt
flask>=2.2.2,<3.0.0
python-dotenv>=0.21.1,<1.0.0
flask_sqlalchemy
flask_migrate>=4.0.4,<5.0.0
flask_login>=0.6.2,<1.0.0
flask_wtf>=1.1.1,<2.0.0
wtforms>=3.0.1,<4.0.0
pillow>=9.4.0,<10.0.0
flask_admin>=1.6.0,<2.0.0
werkzeug